refactor / Add editSvg.ts
