+++ package.json
@@ -33,6 +33,7 @@
 	},
 	"devDependencies": {
 		"@babel/preset-react": "^7.16.7",
+		"@types/dompurify": "^2.3.3",
 		"@wordpress/e2e-test-utils": "^6.0.1",
 		"@wordpress/jest-preset-default": "^8.0.1",
 		"@wordpress/jest-puppeteer-axe": "^4.0.1",
@@ -43,6 +44,7 @@
 		"npm-check-updates": "^12.4.0",
 		"npm-run-all": "^4.1.5",
 		"prettier": "^2.5.1",
+		"typescript": "^4.7.4",
 		"uuid": "^8.3.2"
 	}
 }
